Update ds18b20-node.js
Added a mechanism whereby setting msg.family to a value causes the directory name to change as appropriate to the correct value for the sensor family. The current set of families is "28" for DS18B20, "10" for DS18S20. It makes more sense to search /sys/devices/w1_bus_master1/w1_master_slaves file for the list of devices. However, this should provide a simple temporary solution.